% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_of.R
\name{any_of}
\alias{any_of}
\title{Match any of these characters exactly once.}
\usage{
any_of(.data = NULL, value)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}

\item{value}{Expression to optionally match}
}
\description{
Constructs a \emph{character class}, sometimes called a
\emph{character set}. With this particular expression, you can tell the
regex engine to match only one out of several characters. Simply place the
characters you want to match between square brackets.
}
\examples{
any_of(value = "abc")

# create an expression
x <- any_of(value = "abc")

grepl(x, "c") # should be true
grepl(x, "d") # should be false

}
\references{
Character class: \url{https://www.regular-expressions.info/charclass.html}
}
