% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anything_but.R
\name{anything_but}
\alias{anything_but}
\title{Match any character(s) except these any (including zero) number of times.}
\usage{
anything_but(.data = NULL, value)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}

\item{value}{Characters to not match}
}
\description{
Match any character(s) except these any (including zero) number of times.
}
\details{
This function is slighly different than \code{start_of_line()}
because it has the additional argument \code{.data} which allows the user to
pass a value to to it. This is meant to allow \code{dplyr} chains to
function properly as \code{.data} defaults to \code{NULL} and will pull
the data from the previous step in the chain.
}
\examples{
anything_but(value = "abc")
}
