% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br.R
\name{br}
\alias{br}
\title{Match a line break (alias of \code{line_break()}).}
\usage{
br(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
Match a line break (alias of \code{line_break()}).
}
\details{
This function is looks for line breaks with the following
expression: \code{(?:\\r\\n|\\r|\\n)}
\enumerate{
  \item Unix style: \url{https://codepoints.net/U+000A}
  \item Windows style: \url{https://codepoints.net/U+000D}
}
}
\examples{
br()

# create an expression
x <- br()

# create input
string <- "foo\\nbar"

# extract match
regmatches(string, regexpr(x, string))
}
