% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_of_line.R
\name{end_of_line}
\alias{end_of_line}
\title{Match the expression only if it appears till the end of the line.}
\usage{
end_of_line(.data = NULL, enable = TRUE)
}
\arguments{
\item{.data}{Expression to match, typically pulled from the pipe \code{ \%>\% }}

\item{enable}{Whether to enable this behavior, defaults to \code{TRUE}}
}
\description{
Control whether to match the expression only if it appears till
the end of the line.
}
\details{
This function is slighly different than \code{start_of_line()}
because it has the additional argument \code{.data} which allows the user to
pass a value to to it. This is meant to allow \code{dplyr} chains to
function properly as \code{.data} defaults to \code{NULL} and will pull
the data from the previous step in the chain.
}
\examples{
end_of_line(enable = TRUE)
end_of_line(enable = FALSE)
end_of_line("abc", enable = TRUE)

# create expression
x <- start_of_line(FALSE) \%>\%
  then("apple") \%>\%
  end_of_line()

grepl(x, "apples") # should be false
grepl(x, "apple")  # should be true
}
