% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_break.R
\name{line_break}
\alias{line_break}
\title{Match a line break.}
\usage{
line_break(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
Match a line break.
}
\details{
This function is looks for line breaks with the following
expression: \code{(?:\\r\\n|\\r|\\n)}
\enumerate{
  \item Unix style: \url{https://codepoints.net/U+000A}
  \item Windows style: \url{https://codepoints.net/U+000D}
}
}
\examples{
line_break()

# create an expression
x <- line_break()

# create input
string <- "foo\\nbar"

# extract match
regmatches(string, regexpr(x, string))
}
