% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{maybe}
\alias{maybe}
\title{Optionally match an expression.}
\usage{
maybe(.data = NULL, value)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}

\item{value}{Expression to optionally match}
}
\description{
This expression uses a \emph{quantifier} \code{?} to optionally
match things. Specifically, the question mark makes the preceding token in
the regular expression optional.
}
\examples{
maybe(value = "abc")

# create expression
x <- start_of_line() \%>\%
  maybe("abc") \%>\%
  end_of_line(enable = FALSE)

grepl(x, "xyz") # should be true

}
\references{
Quantifiers: \url{https://www.regular-expressions.info/optional.html}
}
