% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.R
\name{not}
\alias{not}
\title{Ensure that the parameter does not follow.}
\usage{
not(.data = NULL, value)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}

\item{value}{Value to ensure absence of}
}
\description{
This expression uses a \emph{negative lookahead} to ensure the
value given does not follow the previous verbal expression.
}
\examples{
not(value = "FEB-28")

# construct expression
x <- start_of_line() \%>\%
  find('FEB-29') \%>\%
  not("FEB-28")

# create a string
string <- c("FEB-29-2017", "FEB-28-2017")

# extract matches, perl = TRUE is required for negative lookahead
regmatches(string, regexpr(x, string, perl = TRUE))

}
\references{
Negative lookahead: \url{https://www.regular-expressions.info/lookaround.html}
}
