% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{rx_alpha}
\alias{rx_alpha}
\title{Match alphabetic characters.}
\usage{
rx_alpha(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
Matches letters (case insensitive) only.
}
\examples{
rx_alpha()

# create an expression
x <- rx_alpha()

# create input
string <- "Apple 1!"

# extract match
regmatches(string, gregexpr(x, string))
}
