% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alnum.R
\name{rx_alpha_num}
\alias{rx_alpha_num}
\title{Match alphanumeric characters together with underscore (aka. "word character")}
\usage{
rx_alpha_num(.data = NULL, rep = NULL, mode = "greedy",
  negate = FALSE)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
Matches both letters (case insensitive), underscore \code{_} and numbers (a through
z and 0 through 9).
}
\examples{
rx_alpha_num()

# create an expression
x <- rx_alpha_num()

# create input
string <- "Apple_1!"

# extract match
regmatches(string, gregexpr(x, string))
}
