% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anything.R
\name{rx_anything}
\alias{rx_anything}
\title{Match any character(s) any (including zero) number of times.}
\usage{
rx_anything(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
This expression will match everything except line breaks using
the \emph{dot} and the \emph{star}. The Dot \code{.} is a
\emph{metacharacter} and the Star \code{*} is a \emph{quantifier}. When
combined the expression is considered greedy because it will match everything
(except line breaks) 0 or more times.
}
\examples{
rx_anything()

x <- rx_start_of_line() \%>\%
  rx_anything() \%>\%
  rx_end_of_line()

grepl(x, "anything!")  # this should be true
grepl(rx_anything(), "")  # this should be true
grepl(rx_something(), "") # this should be false

}
\references{
Dot: \url{https://www.regular-expressions.info/dot.html}

Star Quantifier: \url{https://www.regular-expressions.info/repeat.html}
}
