% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_groups.R
\name{rx_begin_capture}
\alias{rx_begin_capture}
\title{Begin a capture group.}
\usage{
rx_begin_capture(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
Begin a capture group.
}
\details{
Capture groups are used to extract data from within the regular
expression match for further processing.
}
