% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit.R
\name{rx_digit}
\alias{rx_digit}
\title{Match a digit (0–9).}
\usage{
rx_digit(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
Match a digit (0–9).
}
\details{
This function is looks for tabs with the following expression:
\code{\\d}
}
\examples{
rx_digit()

# create an expression
x <- rx_digit()

# create input
string <- "1 apple"

# extract match
regmatches(string, regexpr(x, string))
}
