% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_groups.R
\name{rx_end_capture}
\alias{rx_end_capture}
\title{End a capture group.}
\usage{
rx_end_capture(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{\%>\%}}
}
\description{
End a capture group.
}
\details{
Capture groups are used to extract data from within the regular
expression match for further processing.
}
