% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowercase.R
\name{rx_lowercase}
\alias{rx_lowercase}
\title{Match lower case letters.}
\usage{
rx_lowercase(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
Matches lower case letters only.
}
\examples{
rx_lowercase()

# create an expression
x <- rx_lowercase()

# create input
string <- "Apple 1!"

# extract match
regmatches(string, gregexpr(x, string))
}
