% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loops.R
\name{rx_multiple}
\alias{rx_multiple}
\title{Match the previous group any number of times.}
\usage{
rx_multiple(.data = NULL, value = NULL, min = NULL, max = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}

\item{value}{Item to match}

\item{min}{Minimum number of times it should be present}

\item{max}{Maximum number of times it should be present}
}
\description{
Match the previous group any number of times.
}
