% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_of.R
\name{rx_one_of}
\alias{rx_one_of}
\title{Match any of these characters exactly once.}
\usage{
rx_one_of(.data = NULL, ..., rep = NULL, mode = "greedy")
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{\%>\%}}

\item{...}{Literal expressions and \code{rx} statements to optionally match}
}
\description{
Constructs a \emph{character class}, sometimes called a
\emph{character set}. With this particular expression, you can tell the
regex engine to match only one out of several characters. It does this by
simply placing the characters you want to match between square brackets.
}
\note{
This function *can not* stand first in the chain of \code{rx} statements and requires constructor function \code{rx()}
when used in piped series of commands.
}
\examples{
rx() \%>\%
   rx_one_of("abc")

# create an expression
x <- rx() \%>\% rx_one_of("abc")

grepl(x, "c") # should be true
grepl(x, "d") # should be false

y <- rx_find("gr") \%>\%
  rx_one_of("ae") \%>\%
  rx_find("y")

regmatches("gray", regexec(y, "gray"))[[1]]
regmatches("grey", regexec(y, "grey"))[[1]]

}
\references{
Character class: \url{https://www.regular-expressions.info/charclass.html}
}
