% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_or_more.R
\name{rx_one_or_more}
\alias{rx_one_or_more}
\title{Match the previous stuff one or more times.}
\usage{
rx_one_or_more(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
This function simply adds a + to the end of the expression.
}
\examples{
rx_one_or_more()

# create an expression
x <- rx_find(value = "a") \%>\%
  rx_one_or_more()

# create input
input <- "aaa"

# extract match
regmatches(input, regexpr(x, input))
}
