% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rx_or}
\alias{rx_or}
\title{Alternatively, match another expression.}
\usage{
rx_or(.data, value)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}

\item{value}{A verbal expression}
}
\description{
Expression to match instead. If both expressions exists, both
will be returned. Note that this expression requires a verbal expression for
the value parameter, see the example below. This may change in the future as
I'd like to avoid nesting a function call in \code{or()}.
}
\examples{
# create an expression
x <- rx_find(value = "foo") \%>\%
  rx_or(rx_find(value = "bar"))

# create strings
string1 <- "foo!"
string2 <- "bar!"

# extract matches
regmatches(string1, gregexpr(x, string1))[[1]]
regmatches(string2, gregexpr(x, string2))[[1]]

}
