% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punctuation.R
\name{rx_punctuation}
\alias{rx_punctuation}
\title{Match punctuation characters.}
\usage{
rx_punctuation(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
Matches punctuation characters only:
\code{! \" # $ \% & ' ( ) * + , - . / : ; < = > ? @ [ \\ ] ^ _ ` { | } ~}.
}
\examples{
rx_punctuation()

# create an expression
x <- rx_punctuation()

# create input
string <- 'Apple 1!'

# extract match
regmatches(string, gregexpr(x, string))
}
