% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{rx_repeat_previous}
\alias{rx_repeat_previous}
\title{Repeat the previous item exactly count times.}
\usage{
rx_repeat_previous(.data, n)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}

\item{n}{Number of times to repeat}
}
\description{
Repeat the previous item exactly count times.
}
