% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seek.R
\name{rx_seek_prefix}
\alias{rx_seek_prefix}
\alias{rx_seek_suffix}
\title{Positive lookaround functions}
\usage{
rx_seek_prefix(.data = NULL, value)

rx_seek_suffix(.data = NULL, value)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}

\item{value}{Exact expression to match}
}
\description{
This function facilitates matching by providing assurances for surrounding symbols/groups of symbols.
It allows for building expressions that are dependent on context of occurence.
}
\examples{
# this will match anything between square brackets
rx() \%>\%
  rx_seek_prefix("[") \%>\%
  rx_anything("lazy") \%>\%
  rx_seek_suffix(']')

}
