% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{rx_space}
\alias{rx_space}
\title{Match a regular space character.}
\usage{
rx_space(.data = NULL, rep = NULL, mode = "greedy", negate = FALSE)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{\%>\%}}

\item{negate}{Invert match behavior, defaults to \code{FALSE} (match
space). Use \code{FALSE} to \emph{not} match space.}
}
\description{
Matches a literal space character.
}
\examples{
# match space, default
rx_space()

# dont match space
rx_space(negate = TRUE)

# create an expression
x <- rx_space()

# create input
string <- "1 apple\\t"

# extract match
regmatches(string, regexpr(x, string))

# extract no whitespace by inverting behavior
y <- rx_space(negate = TRUE)
regmatches(string, gregexpr(y, string))
}
