% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{rx_test}
\alias{rx_test}
\title{Test rx expression.}
\usage{
rx_test(x, txt)
}
\arguments{
\item{x}{\code{rx} expression to test}

\item{txt}{Text string to be used for matching a pattern.}
}
\description{
The function \code{rx_test()} is intended for quick testing of \code{rx} expressions without loading new packages.
It uses combination of \code{regexmatches} and \code{gregexpr} to extract all matches of pattern in a string. Since
\code{rx} pattern argument is first, it support piping and can be placed at the end of the \code{rx} pipe for testing purposes.
}
\examples{
rx_digit() \%>\%
  rx_test("Why is 6 afraid of 7? Because 7,8,9")
}
