% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper.R
\name{rx_upper}
\alias{rx_upper}
\title{Match upper case letters.}
\usage{
rx_upper(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
Matches upper case letters only.
}
\examples{
rx_upper()

# create an expression
x <- rx_upper()

# create input
string <- "Apple 1!"

# extract match
regmatches(string, gregexpr(x, string))
}
