% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uppercase.R
\name{rx_uppercase}
\alias{rx_uppercase}
\title{Match upper case letters.}
\usage{
rx_uppercase(.data = NULL, rep = NULL, mode = "greedy",
  negate = FALSE)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
Matches upper case letters only.
}
\examples{
rx_uppercase()

# create an expression
x <- rx_uppercase()

# create input
string <- "Apple 1!"

# extract match
regmatches(string, gregexpr(x, string))
}
