% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{rx_word}
\alias{rx_word}
\title{Match a word.}
\usage{
rx_word(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
Match a word—a string of word characters (a–z, A–Z, 0–9 or _).
This function is looks for tabs with the following expression: \code{\\w+}
}
\examples{
rx_word()

# create an expression
x <- rx_word()

# create inputs
string1 <- "foo_bar"
string2 <- "foo-bar"

# extract matches
regmatches(string1, regexpr(x, string1))
regmatches(string2, regexpr(x, string2)) # doesn't match -
}
