% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_characters.R
\name{rx_word_char}
\alias{rx_word_char}
\title{Match a word character.}
\usage{
rx_word_char(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
Match a word character (a–z, A–Z, 0–9 or _).
}
\examples{
rx_word_char()

# Same as rx_word()
x <- rx_word_char() \%>\%
 rx_one_or_more()

}
