% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rx_word_edge}
\alias{rx_word_edge}
\title{Find beginning or end of a word.}
\usage{
rx_word_edge(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{\%>\%}}
}
\description{
Match beginning or end of a word—a string consisting of of word
characters (a–z, A–Z, 0–9 or _).
}
\examples{
rx_word_edge()

x <- rx() \%>\%
 rx_word_edge() \%>\%
 rx_alpha() \%>\%
 rx_one_or_more() \%>\%
 rx_word_edge()

# create inputs
string1 <- "foobar"
string2 <- "foo 23a bar"

# matches 'foobar'
regmatches(string1, regexpr(x, string1))
# matches 'foo' and 'bar' separately
regmatches(string2, gregexpr(x, string2))
}
