% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize}
\alias{sanitize}
\title{Escape characters expected special by regex engines}
\usage{
sanitize(.data = NULL)
}
\arguments{
\item{.data}{String to sanitize}
}
\description{
Takes a string and escapes all characters considered special by
the regex engine. This is used internally when you add a string to the
\code{.data} parameter in most of the available functions. It is exported
and useable externally for users that want to escape all special characters
in their desired match. The following special characters are escaped
\code{. | * ? + ( ) { } ^ $ \\ : = [ ]}
}
\examples{
sanitize("^")
sanitize("^+")
sanitize("^+?")
}
