% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/something_but.R
\name{something_but}
\alias{something_but}
\title{Match any character(s) except these at least once.}
\usage{
something_but(.data = NULL, value)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}

\item{value}{Expression to optionally match}
}
\description{
Match any character(s) except these at least once.
}
\details{
This function is slighly different than \code{start_of_line()}
because it has the additional argument \code{.data} which allows the user to
pass a value to to it. This is meant to allow \code{dplyr} chains to
function properly as \code{.data} defaults to \code{NULL} and will pull
the data from the previous step in the chain.
}
\examples{
something_but(value = "abc")

# create an expression
x <- something_but(value = "python")

grepl(x, "R")  # should be true
grepl(x, "py") # should be false
}
