% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_of_line.R
\name{start_of_line}
\alias{start_of_line}
\title{Match the expression only if it appears from beginning of line.}
\usage{
start_of_line(enable = TRUE)
}
\arguments{
\item{enable}{Whether to enable this behavior, defaults to \code{TRUE}}
}
\description{
Control whether to match the expression only if it appears from
the beginning of the line.
}
\examples{
start_of_line(enable = TRUE)
start_of_line(enable = FALSE)

# create expression
x <- start_of_line() \%>\%
  then("apple")

grepl(x, "pineapple") # should be false
grepl(x, "apple")     # should be true
}
