% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Match a tab character.}
\usage{
tab(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
Match a tab character.
}
\details{
This function is looks for tabs with the following
expression: \code{\\t}
\enumerate{
  \item Tab character: \url{https://codepoints.net/U+0009}
}
}
\examples{
tab()

# create an expression
x <- tab()

# create input
string <- "foo\\tbar"

# extract match
regmatches(string, regexpr(x, string))
}
