% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_of.R
\name{any_of}
\alias{any_of}
\title{Match any of these characters exactly once.}
\usage{
any_of(.data = NULL, value)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}

\item{value}{Expression to optionally match}
}
\description{
Match any of these characters exactly once.
}
\details{
This function is slighly different than \code{start_of_line()}
because it has the additional argument \code{.data} which allows the user to
pass a value to to it. This is meant to allow \code{dplyr} chains to
function properly as \code{.data} defaults to \code{NULL} and will pull
the data from the previous step in the chain.
}
\examples{
any_of(value = "abc")

# create an expression
x <- any_of(value = "abc")

grepl(x, "c") # should be true
grepl(x, "d") # should be false
}
