% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anything.R
\name{anything}
\alias{anything}
\title{Match any character(s) any (including zero) number of times.}
\usage{
anything(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
This expression will match everything except line breaks using
the \emph{dot} and the \emph{star}. The Dot \code{.} is a
\emph{metacharacter} and the Star \code{*} is a \emph{quantifier}. When
combined the expression is considered greedy because it will match everything
(except line breaks) 0 or more times.
}
\examples{
anything()

x <- start_of_line() \%>\%
  anything() \%>\%
  end_of_line()

grepl(x, "anything!") # this should be true

}
\references{
Dot: \url{https://www.regular-expressions.info/dot.html}

Star Quantifier: \url{https://www.regular-expressions.info/repeat.html}
}
