% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br.R
\name{br}
\alias{br}
\title{Match a line break (alias of \code{line_break()}).}
\usage{
br(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
This expression looks for line breaks, both Unix and Windows
style by using the appropriate \emph{non printable characters}.
}
\examples{
br()

# create an expression
x <- br()

# create input
string <- "foo\\nbar"

# extract match
regmatches(string, regexpr(x, string))

}
\references{
Unix style: \url{https://codepoints.net/U+000A}

Windows style: \url{https://codepoints.net/U+000D}

Non printable character: \url{https://www.regular-expressions.info/nonprint.html}
}
