% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{maybe}
\alias{maybe}
\title{Optionally match an expression.}
\usage{
maybe(.data = NULL, value)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}

\item{value}{Expression to optionally match}
}
\description{
Optionally match an expression.
}
\details{
This function is slighly different than \code{start_of_line()}
because it has the additional argument \code{.data} which allows the user to
pass a value to to it. This is meant to allow \code{dplyr} chains to
function properly as \code{.data} defaults to \code{NULL} and will pull
the data from the previous step in the chain.
}
\examples{
maybe(value = "abc")

# create expression
x <- start_of_line() \%>\%
  maybe("abc") \%>\%
  end_of_line(enable = FALSE)

grepl(x, "xyz") # should be true
}
