% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{rx_count}
\alias{rx_count}
\title{Match the previous stuff exact number of times.}
\usage{
rx_count(.data = NULL, n = 1)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}

\item{n}{Number of times previous expression shall be repeated. Default is 1.}
}
\description{
This function simply adds a \code{{n}} to the end of the expression.
}
\examples{
rx_count()

# create an expression
x <- rx_find(value = "a") \%>\%
  rx_count(3)

# create input
input <- "aaa"

# extract match
regmatches(input, regexpr(x, input))

}
