% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitespace.R
\name{rx_whitespace}
\alias{rx_whitespace}
\title{Match a whitespace character.}
\usage{
rx_whitespace(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
Match a whitespace character.
}
\details{
Match a whitespace character (one of space, tab, carriage return,
new line, vertical tab and form feed). This function is looks for tabs with
the following expression: \code{\\d}
\enumerate{
  \item space: \url{https://codepoints.net/U+0020}
  \item tab: \url{https://codepoints.net/U+0009}
  \item carriage return: \url{https://codepoints.net/U+000D}
  \item new line: \url{https://codepoints.net/U+000}
  \item vertical tab: \url{https://codepoints.net/U+000B}
  \item form feed: \url{https://codepoints.net/U+000C}
}
}
\examples{
rx_whitespace()

# create an expression
x <- rx_whitespace()

# create input
string <- "1 apple"

# extract match
regmatches(string, regexpr(x, string))
}
