% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/something.R
\name{something}
\alias{something}
\title{Match any character(s) at least once.}
\usage{
something(.data = NULL)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}
}
\description{
This expression is almost identical to \code{anything()}
with one major exception, a \code{+} is used instead of a \code{*}. This
means \code{something()} expects \emph{something} whereas
\code{anything()} expects \emph{anything} including... nothing!
}
\examples{
something()

# construct an expression
x <- something()

grepl(x, "something!") # this should be true
grepl(x, "")           # this should be false
grepl(anything(), "")  # this should be true

}
\references{
Metacharacters: \url{https://www.regular-expressions.info/characters.html#special}
}
