% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/something_but.R
\name{something_but}
\alias{something_but}
\title{Match any character(s) except these at least once.}
\usage{
something_but(.data = NULL, value)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}

\item{value}{Expression to optionally match}
}
\description{
This expression is almost identical to \code{anything_but()}
with one major exception, a \code{+} is used instead of a \code{*}. This
means \code{something_but()} expects \emph{something} whereas
\code{anything_but()} expects \emph{anything} including... nothing!
}
\examples{
something_but(value = "abc")

# create an expression
x <- something_but(value = "python")

grepl(x, "R")  # should be true
grepl(x, "py") # should be false

}
\references{
Metacharacters: \url{https://www.regular-expressions.info/characters.html#special}
}
