% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/then.R
\name{then}
\alias{then}
\title{Match an expression}
\usage{
then(.data = NULL, value)
}
\arguments{
\item{.data}{Expression to append, typically pulled from the pipe \code{ \%>\% }}

\item{value}{Expression to match}
}
\description{
This expression uses a \emph{non capturing group} to identify a
specific pattern when you do not need the group to capture its match. Alias
for then. Meant for semantics when used at the beginning of a verbal
expression. For example, \code{find(value = 'foo')} is more readable than
\code{then(value = 'foo')}.
}
\examples{
then(value = "apple")

# create expression
x <- start_of_line() \%>\%
  then("apple") \%>\%
  end_of_line()

grepl(x, "apples") # should be false
grepl(x, "apple")  # should be true

}
\references{
Non capturing group: \url{https://www.regular-expressions.info/brackets.html}

Stack Overflow: \url{https://stackoverflow.com/questions/3512471}
}
